Modularization completed - minor bug - sometimes bot doesnt start, just keep on trying
