+++ bot.js
@@ -1,5 +1,6 @@
 const { bot } = require('./createBot.js')
 const { sayHelloOnUserArrival, sayHelloOnBotLogIn } = require('./greetings')
+const { processBotCommand, checkMessageForBotCommand } = require('./processCommands')
 
 
 bot.on('start', function() {
@@ -13,7 +14,7 @@ bot.on('message', (event) => {
     console.log(event)
 
 
-    // check if user is logging in to slack
+    // check if user is logging in to slack and send them a hello message
     if(event.type === 'presence_change' && event.presence === 'active') {
         sayHelloOnUserArrival(event.user)
     }
@@ -32,62 +33,3 @@ bot.on('message', (event) => {
     }
 
 })
-
-
-const checkMessageForBotCommand = (channelText) => {
-    // console.log("text", channelText)
-
-    // "!say whatever"
-
-    let commandObj = {
-        type: null,
-        argument: null
-    }
-
-    let channelTextArray = channelText.split(' ')
-
-    let command = channelText.split(' ')[0] // ['!say', 'whatever']
-
-    let argument = channelTextArray.slice(1, channelTextArray.length).join(' ');
-
-    // console.log("checkForCommand", channelTextArray[0])
-    // console.log("argument", argument)
-
-    if(command[0] === "!") {
-        commandObj.type = command.slice(1);
-        console.log("type", commandObj.type)
-        commandObj.argument = argument;
-
-        // console.log("command", commandObj)
-
-        return commandObj
-    } else {
-        return false
-    }
-}
-
-
-const processBotCommand = (command) => {
-    console.log("process command test", command)
-
-    let params = {
-        icon_emoji: ':robot_face:'
-    };
-
-    // SWITCH STATEMENT WAY THAT WORKS
-
-    switch(command.type){
-
-        case 'say':
-            bot.postTo('general', command.argument, params);
-            break;
-
-        case 'killbot':
-            process.exit()
-            break;
-
-        default:
-            console.log("Improper command entered");
-
-    }
-}
