+++ greetings.js
@@ -3,6 +3,9 @@ const { bot } = require('./createBot')
 
 // set params for robot for all greet messages
 // more information about additional params https://api.slack.com/methods/chat.postMessage
+
+  // Params are set locally because of a strange bug when trying to import from the createBot module
+
 let params = {
       icon_emoji: ':robot_face:'
   };
