+++ tictactoe.js
@@ -1,0 +1,8 @@
+const { bot } = require('./createBot')
+
+
+let createNewTicTacToeGame = () => {
+  console.log("Creating new TTT Game")
+}
+
+module.exports = { createNewTicTacToeGame }
